Moto Guzzi jiffy warning signal on dashboard
My Moto Guzzi will stall if the jiffy is out. Unfortunately, this also
happens when the jiffy switch is broke or dirty (e.g. because of too
much oil/grease usedd on the jiffy). This can be annoying and even
dangerous. This project is about disabling the automatic inactivation of
the motorcycle, and instead of this, only signalling it with a led on
the dashboard if the jiffy is out.
Some goals: use both sides of the jiffy switch for increased
reliability, switch the signalling led on the +12V side, so the other
side can be connected to ground (like all other bulbs in the motor
cycle), make the signalling led flash if the jiffy is out.